fix circular dep between types and mutate-request
