+++ src/lib/adjudicators.test.ts
@@ -4,7 +4,8 @@
 import { expect, describe, it } from "@jest/globals";
 import * as sut from "./adjudicators";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { AdmissionRequest, Binding, DeepPartial, Event, Operation } from "./types";
+import { AdmissionRequest, Binding, Event } from "./types";
+import { DeepPartial, Operation } from "./mutate-types";
 
 describe("definesDeletionTimestamp", () => {
   //[ Binding, result ]
