+++ src/lib/capability.test.ts
@@ -4,9 +4,9 @@ import { CapabilityCfg, FinalizeAction, MutateAction, ValidateAction, WatchLogAc
 import { a } from "../lib";
 import { V1Pod } from "@kubernetes/client-node";
 import { expect, describe, jest, beforeEach, it } from "@jest/globals";
-import { PeprMutateRequest } from "./mutate-request";
+import { PeprMutateRequest, Operation } from "./mutate-types";
 import { PeprValidateRequest } from "./validate-request";
-import { Operation, AdmissionRequest } from "./types";
+import { AdmissionRequest } from "./types";
 import { WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
 import { Event } from "./types";
 import { GenericClass } from "kubernetes-fluent-client";
