+++ src/lib/finalizer.ts
@@ -3,9 +3,9 @@
 
 import { K8s, KubernetesObject, RegisterKind } from "kubernetes-fluent-client";
 import Log from "./logger";
-import { Binding, DeepPartial } from "./types";
-import { Operation } from "./types";
-import { PeprMutateRequest } from "./mutate-request";
+import { Binding } from "./types";
+import { Operation, DeepPartial } from "./mutate-types";
+import { PeprMutateRequest } from "./mutate-types";
 
 export function addFinalizer<K extends KubernetesObject>(request: PeprMutateRequest<K>) {
   // if a DELETE is being processed, don't add a finalizer
