+++ src/lib/types.ts
@@ -2,20 +2,21 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
 import { GenericClass, GroupVersionKind, KubernetesObject } from "kubernetes-fluent-client";
-
+import { Operation } from "./mutate-types";
 import { WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
-
-import { PeprMutateRequest } from "./mutate-request";
-import { PeprValidateRequest } from "./validate-request";
-
 import { Logger } from "pino";
 
-export enum Operation {
+import { PeprMutateRequest } from "./mutate-types";
+import { PeprValidateRequest } from "./validate-request";
+
+//import { Logger } from "pino";
+
+/* export enum Operation {
   CREATE = "CREATE",
   UPDATE = "UPDATE",
   DELETE = "DELETE",
   CONNECT = "CONNECT",
-}
+} */
 /**
  * Specifically for deploying images with a private registry
  */
@@ -43,9 +44,9 @@ export interface ResponseItem {
 /**
  * Recursively make all properties in T optional.
  */
-export type DeepPartial<T> = {
+/* export type DeepPartial<T> = {
   [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
-};
+}; */
 
 /**
  * The type of Kubernetes mutating webhook event that the action is registered for.
