+++ src/lib/validate-request.test.ts
@@ -3,7 +3,8 @@
 
 import { beforeEach, describe, expect, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { ValidateActionResponse, AdmissionRequest, Operation } from "./types";
+import { ValidateActionResponse, AdmissionRequest } from "./types";
+import { Operation } from "./mutate-types";
 import { PeprValidateRequest } from "./validate-request";
 describe("PeprValidateRequest", () => {
   let mockRequest: AdmissionRequest<KubernetesObject>;
